Inline the single Debian non-APT step
