+++ .debride-whitelist
@@ -2,7 +2,6 @@
 /^on_/
 config=
 down
-install_direct_download
 log_file
 print_steps
 run_if_existing_machine
