+++ lib/dotfiles/steps/debian/install_debian_droid_step.rb
@@ -1,25 +1,51 @@
 class Dotfiles::Step::InstallDebianDroidStep < Dotfiles::Step
   DESCRIPTION = "Installs Factory Droid on Debian/Ubuntu when configured as a non-APT package.".freeze
-
-  include Dotfiles::Step::DebianNonAptStep
-
   DROID_INSTALL_URL = "https://app.factory.ai/cli".freeze
+
+  debian_only
 
   def self.display_name
     "Factory Droid"
   end
 
+  def should_run?
+    allowed_on_platform? && configured? && !installed?
+  end
+
+  def run
+    install if configured?
+  end
+
+  def complete?
+    super
+    return true unless configured?
+    return true if installed?
+
+    add_error("Non-APT package not installed: droid")
+    false
+  end
+
   private
 
-  def package_name
-    "droid"
+  def configured?
+    @config.debian_non_apt_packages.include?("droid")
   end
 
-  def install_script_url
-    DROID_INSTALL_URL
+  def installed?
+    command_exists?("droid") || droid_bin_paths.any? { |path| @system.file_exist?(path) }
   end
 
-  def install_error_label
-    "Droid"
+  def droid_bin_paths
+    [
+      File.join(@home, ".local", "bin", "droid"),
+      File.join(@home, ".cargo", "bin", "droid")
+    ]
+  end
+
+  def install
+    return if installed?
+
+    output, status = execute(shell_script('curl -fsSL "$1" | "$2"', DROID_INSTALL_URL, "sh"))
+    add_error("Droid install failed (status #{status}): #{output}") unless status == 0
   end
 end
